% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.p.bdp.R
\name{alpha.p.bdp}
\alias{alpha.p.bdp}
\title{Computes \eqn{\alpha}, power of DPD}
\usage{
alpha.p.bdp(p, bdp)
}
\arguments{
\item{p}{numeric; number of variables.}

\item{bdp}{numeric; nominal breakdown point.}
}
\value{
numeric scalar
}
\description{
Computes \eqn{\alpha} for DPD given dimension \emph{p} and nominal breakdown point \emph{bdp}
}
\examples{
alpha = alpha.p.bdp(p = 5, bdp = 0.12)
alpha
alpha = alpha.p.bdp(p = 10, bdp = 0.03)
alpha
}
\seealso{
\code{\link{covMdpd}}, \code{\link{bdp.max}}, \code{\link{bdp.p.alpha}}, \code{\link{h.bdp.n.p}}
}
