% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdp.max.R
\name{bdp.max}
\alias{bdp.max}
\title{Computes maximum nominal breakdown point for DPD}
\usage{
bdp.max(p)
}
\arguments{
\item{p}{numeric; number of variables.}
}
\value{
numeric scalar
}
\description{
Computes the maximum nominal breakdown point for DPD given \emph{p}.
}
\examples{
bdp.max(p = 5)
bdp.max(p = 15)
}
\seealso{
\code{\link{covMdpd}}, \code{\link{bdp.p.alpha}}, \code{\link{alpha.p.bdp}}, \code{\link{h.bdp.n.p}}
}
