% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdp.p.alpha.R
\name{bdp.p.alpha}
\alias{bdp.p.alpha}
\title{Computes nominal breakdown point for DPD}
\usage{
bdp.p.alpha(p, alpha)
}
\arguments{
\item{p}{numeric; number of variables.}

\item{alpha}{numeric value between 0 and 1; represents DPD power.}
}
\value{
numeric scalar
}
\description{
Computes the nominal breakdown point for DPD given \emph{p} and \eqn{\alpha}
}
\examples{
bdp.p.alpha(p = 5, alpha = 0.5)
bdp.p.alpha(p = 8, alpha = 0.75)
}
\seealso{
\code{\link{covMdpd}}, \code{\link{bdp.max}}, \code{\link{alpha.p.bdp}}, \code{\link{h.bdp.n.p}}
}
