% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpd.R
\name{covMdpd}
\alias{covMdpd}
\title{Robust Estimation of Mean Vector and Covariance Matrix for Contaminated Gaussian Datasets via Density Power Divergence Minimization}
\usage{
covMdpd(x)
covMdpd(x, alpha = 2.0/ncol(x), h.min = h.bdp.n.p(0.5, nrow(x), ncol(x)), rcond.min = 1E-6, tolIter = 1E-9, maxiter = 500L)

## S3 method for class 'mdpd'
print.mdpd(x, digits = getOption("digits"), prefix = "\t",...)

## S3 method for class 'mdpd'
plot.mdpd(x, which = c("all", "distance", "qqchi2"), classic = FALSE)
}
\arguments{
\item{x}{multivariate data set; matrix.}

\item{alpha}{numeric value between 0 and 1; represents DPD power.}

\item{h.min}{numeric; represents the smallest desired best subset size.}

\item{rcond.min}{numeric; condition number tolerance.}

\item{tolIter}{numeric; iteration error tolerance.}

\item{maxiter}{numeric; maximum iteration number. Default value set to 500.}
}
\value{
A list with components
\item{raw.center }{ Location vector}
\item{raw.cov }{ Scatter matrix}
\item{raw.weights }{ Weights}
\item{raw.mah }{ Squared Mahalanobis distances}
\item{crit }{ DPD objective value}
\item{iter }{ Number of iterations}
}
\description{
Computes MDPD estimator via custom M-iteration
}
\details{
Consider the \emph{p}-variate DPD objective function

\deqn{H_n(\boldsymbol{\theta}) = \int_{\mathbb{R}^{p}} f^{1+\alpha}(\boldsymbol{x}|\boldsymbol{\theta}) \mathrm{d}\boldsymbol{x} -
\left(1 + \frac{1}{\alpha} \right) n^{-1} \sum_{i=1}^{n} f^{\alpha}(\boldsymbol{x}_{i}|\boldsymbol{\theta}),}
where \eqn{f(\boldsymbol{x}|\boldsymbol{\theta})} is a \emph{p}-variate probability density.
Here, we choose the Gaussian density
\eqn{f(\boldsymbol{x}|\boldsymbol{\mu}, \boldsymbol{\Sigma}) =  (2\pi)^{-\frac{p}{2}} |\boldsymbol{\Sigma}|^{-\frac{1}{2}}
\exp(-\frac{1}{2} \big\|\boldsymbol{\Sigma}^{-1/2}(\boldsymbol{x} - \boldsymbol{\mu})\big\|^{2})} with
\eqn{\boldsymbol{\theta} = (\boldsymbol{\mu}, \boldsymbol{\Sigma})}, where
\eqn{\boldsymbol{\mu} \in \mathbb{R}^{p}} and \eqn{\boldsymbol{\Sigma} \in \mathrm{SPD}(p \times p)} are the location vector and
scatter matrix respectively.
For a given sample \eqn{ \boldsymbol{X} = \{x_1, \ldots, x_n\} } from \eqn{ (1- \epsilon)f(\boldsymbol{x}|\boldsymbol{\theta}) +
\epsilon \mathrm{d}G(\boldsymbol{x}) } with \eqn{\epsilon \in [0, 1/2) } and arbitrary unknown cdf \eqn{G(\boldsymbol{x})},
the minimum density power divergence estimator is given as:
\deqn{\boldsymbol{\hat\theta}^{\mathrm{DPD}}_{n} := \underset{\boldsymbol{\theta} \in \Theta}{\arg \min} \quad H_{n}(\boldsymbol{\theta})}
}
\examples{
n = 100
p = 5
x = MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = diag(p))
covMdpd(x) #prints out a list of outputs including the location vector and scatter matrix

n = 250
p = 10
x = MASS::mvrnorm(n = n, mu = rep(0,p), Sigma = diag(p))
res = covMdpd(x, alpha = 0.5)
print(res)
plot(res, which = "distance", classic = TRUE)

## example with outliers
set.seed(1)
eps = 0.1
n = 250
p = 5
ncp = 150
n.out = eps*n
out.index = sample(x = n, n.out)
x.cont = MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = diag(p))
x.cont[out.index, ] = x.cont[out.index, ] + sqrt(ncp/p)
res = covMdpd(x.cont)
print(res)
plot(res, which = "all", classic = TRUE)
}
\references{
Basu, A., Harris, I. R., Hjort, N. L., and Jones, M. (1998). Robust and efficient estimation by minimising a density power divergence. \emph{Biometrika}, \strong{85}(3):549–559.

Hubert M., Debruyne M., Rousseeuw P.J. (2018). Minimum covariance determinant and extensions. \emph{WIREs Computational Statistics}, 10:e1421.

Rousseeuw, P. J., and van Driessen, K. (1999). A fast algorithm for the minimum covariance determinant estimator. \emph{Technometrics}, \strong{41}, 212–223.
}
\seealso{
\code{\link{covMcd}}, \code{\link{bdp.max}}, \code{\link{bdp.p.alpha}}, \code{\link{alpha.p.bdp}}, \code{\link{h.bdp.n.p}}
}
\author{
Michael Pokojovy (\email{michael.pokojovy@gmail.com}), Andrews T. Anum, Ebenezer Nkum, Abhijit Mandal
}
\keyword{Provide}
\keyword{here}
\keyword{keywords}
