% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.bdp.n.p.R
\name{h.bdp.n.p}
\alias{h.bdp.n.p}
\title{Computes smallest necessary subsample size \emph{h} for multivariate robust estimators}
\usage{
h.bdp.n.p(bdp, n, p)
}
\arguments{
\item{bdp}{numeric value between 0 and 0.5; nominal breakdown point.}

\item{n}{numeric; sample size.}

\item{p}{numeric; number of columns.}
}
\value{
numeric scalar
}
\description{
Computes the smallest necessary uncontaminated subsample size to ensure desired bdp
for given sample size \emph{n} and number of variable \emph{p}
}
\examples{
h.bdp.n.p(bdp = 0.01, n = 100, p = 20)
h.bdp.n.p(bdp = 0.18, n = 50, p = 5)
}
\seealso{
\code{\link{covMdpd}}, \code{\link{h.alpha.n}}, \code{\link{bdp.max}}, \code{\link{bdp.p.alpha}}, \code{\link{alpha.p.bdp}}
}
